+++ hatch.toml
@@ -557,6 +557,7 @@ dependencies = [
     "pytest-cov",
     "requests",
     "hypothesis",
+    "flask",
 ]
 
 [envs.ddtrace_unit_tests.env-vars]
@@ -568,7 +569,8 @@ CMAKE_BUILD_PARALLEL_LEVEL="6"
 test = [
     "uname -a",
     "pip freeze",
-    "python -m pytest tests/internal/service_name/test_extra_services_names.py -vvv -s",
+    "python -m pytest tests/internal/service_name/test_extra_services_names.py -v -s",
+    "python -m pytest tests/appsec/architectures/test_appsec_loading_modules.py -v -s",
 ]
 
 [[envs.ddtrace_unit_tests.matrix]]
